+++ pkg/config/integrations.go
@@ -114,6 +114,22 @@ func (c *VersionedIntegrations) setVersion(v integrationsVersion) error {
 	}
 }
 
+// GetIntegrationsList returns a slice of enabled integrations
+func (c *VersionedIntegrations) GetIntegrationsList() []string {
+	integrations := []string{}
+	if c.configV1 != nil {
+		for _, integration := range c.configV1.Integrations {
+			integrations = append(integrations, integration.Name())
+		}
+	}
+	if c.configV2 != nil {
+		for _, integration := range c.configV2.Configs {
+			integrations = append(integrations, integration.Name())
+		}
+	}
+	return integrations
+}
+
 // IntegrationsGlobals is a global struct shared across integrations.
 type IntegrationsGlobals = v2.Globals